requirements.txt
streamlit
pandas
torch
transformers
scikit-learn
PyPDF2
python-docx
langchain
langchain-openai
openai
